packages/pipeline/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcl-pipeline"
version = "0.1.0"
description = "Influence Pirates content pipeline - core library"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "requests>=2.25.0",
    "yt-dlp>=2023.0.0",
    "python-dotenv>=1.0.0",
    "redis>=5.0.0",
    "supabase>=2.0.0",
    "reportlab>=4.0.0",
    "anthropic>=0.18",
    "google-generativeai>=0.4",
    "openai>=1.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]
instagram = ["instaloader>=4.10"]
whisper = ["openai-whisper>=20231117"]

[tool.hatch.build.targets.wheel]
packages = ["mcl_pipeline"]
